app/requirements.txt
asgiref==3.4.1
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.21.1
drf-yasg2==1.19.4
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.2
pyparsing==3.0.6
pyrsistent==0.18.1
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
